Transition from exception-raising Unix.getenv to Sys.getenv_opt with
explicit handling of failure cases.

OCaml's stdlib has Sys.getenv_opt since 4.05. Some of the newer code
already uses it, and some of the old code handled exceptions (so
could nicely be transitioned to handling options instead). Some,
however, did not handle failure at all. This commit remedies that.

In most cases, getenv is used to query the PATH variable (before
adding another directory to it, for example), in which case there is
a nice default value of "". In some cases, the environment variable
is required to be present to proceed, then there is a failure of some
kind raised with the appropriate message.

A test case was added to the quality-gate.sh script to prevent
introduction of the exception-raising Unix.getenv into new code.

Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>