context: catch error inside span
Update `with_tracing` to record the error and finish the span if the
instrumented function fails.

Add `session.track.id` and `task.id` attributes to all spans inside a
trace that is following a task.

This will result in spans being finished even if the function has raised
an exception. Plus, better debuggability in the case that something
fails.

Signed-off-by: Gabriel Buica <danutgabriel.buica@cloud.com>