CA-390277: Stop using host records on CLI cross-pool migrations
Host records have changed between CH8.2CU1 and XS8, this means that the
records produced from an new host cannot be serialized by this code.
Instead of making the remote host filter records, now the client tries
to fetch the correct reference, and if it can't, fetches all references
and fetches fields to find it.

Records from Networks, PBDs and SRs are still used.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>