Add new check for new parameters' default value
Note we are only checking for default of new parameters in messages
that already have default parameters.

It lacks checks for new parameters added to messages with no default
parameters, but there is no easy way to achieve that because the generator
cannot tell what is new and what is not, moreover, there are existing
messages with non-default parameters introduced in different releases.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>