CP-50078: Check if SMAPI observation is enabled
Modifies xapi-storage-script to determine if SMAPI observation is enabled by
checking whether a configuration file exists for it within the file system.

Internally, xapi queries the database to determine if a specific component
should be observed but as xapi-storage-script is an external daemon, we rely
on the file system.

To avoid modifying much of the monadic structure, a configuration record is
used to record whether observation is enabled. It is hoped this less invasive
change will make it simpler to rework xapi-storage-script in the future.

The path of the observer configuration directory is relocated to xapi-consts.

Signed-off-by: Colin James <colin.barr@cloud.com>