CP-50078: Augment RPC transport to use observer.py
Modifies the RPC transport (fork_exec_rpc) to take a "dbg" parameter that can
encode a traceparent. If observation is enabled and a traceparent can be parsed
from the dbg parameter, the storage plugin's script will be invoked via
observer.py (for automatic instrumentation).

This change does not yet seed each of the transports with the dbg parameter;
the volume, datapath, etc. interfaces - defined in xapi-storage - use the
convention of having dbg as the first parameter for each RPC function.

The act of supplying this dbg parameter to the transport could be done
implicitly, but this would be a contract that is hard to enforce, despite
vastly reducing the pervasiveness of changes to come; each handler will need to
partially apply its received dbg parameter to the RPC transport (e.g.
volume_rpc ~dbg) at each usage site.

The python3 path is relocated from xapi_globs.ml to become a constant within
xapi-consts.

Signed-off-by: Colin James <colin.barr@cloud.com>