CP-50078: Augment RPC transports for tracing
This change is rather pervasive but it seeds the volume and datapath
RPC transport with the received "dbg" parameter at each usage site.

The change could be done implicitly if we could be promised that "dbg" always
occurs as a first parameter - we would then just modify the transport to be
aware of this contract and not have to partially apply it at each usage site.

Signed-off-by: Colin James <colin.barr@cloud.com>