CA-395174: Try to unarchive VM's metrics when they aren't running
Backport 71c39605b3a2dd6eddeba42a5b482b4fc1b3a4e8

Non-running VMs' metrics are stored in the coordinator. When the coordinator is
asked about the metrics try to unarchive them instead of failing while trying
to fetch the coordinator's IP address.

This needs to force the HTTP method of the query to be POST

Also returns a Service Unavailable when the host is marked as Broken.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>
Signed-off-by: Christian Lindig <christian.lindig@cloud.com>