CP-49073: Add multihoming support for cluster
Corosync 3 introduces multihoming support, where a node can be assigned
multiple IP addresses, even with different IP stacks such as IPv4 and
IPv6. Expose this feature in the xapi datamodel by introducing
`other_networks` and `other_pifs`.

The previous `network` and `pif` field remains in place and the extra IPs
will only be used as a fail-over: when corosync fails to communicate through
the main pif, these secondary pifs will be used. In the future we could add
more configs to allow other behaviours of the redundant pifs.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>