Clock: Use integers for date's timezone offsets
Previously the comments and code talked about "timezone printing", but that is
only a consequence of a more fundamental problem: datetimes without timezone
do not share a frame of reference with datetimes that have a timezone, nor
among themselves. This means that timezoneless shouldn't be compared with other
timezones as assuming a timezone can be incorrect.
Datetimes with timezone all share a frame of reference: the unix epoch.

Now the code uses the exact offset in seconds, which also enables it to accept
datetimes that have an offset different than 0.

Comparisons with timezoneless datetimes are not forbidden for the time being.
Instead UTC is assumed, like before, even if it might be the wrong guess.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>