CP-50270: Set the correct parent in `make_connection`
Updates the correct parent span when we retrieve a cached connection.

Preveusly, we would call the instrumentation only when we make the
connection for the first time. This resulted in seeing the same parent
span for all subsequent calls of `make_connection`.
Calling the instrumentation first ensures the correct parent is sent
over the headers for all code branches.

Signed-off-by: Gabriel Buica <danutgabriel.buica@cloud.com>