CA-395784: fix(buf_io_test): the timeout is per read, not per function call
really_input takes a timeout parameter, however this is the timeout
for each individual read call, not the entire function.

So if we have a sender that sends a 64KiB buffer 1 byte at a time,
and we set a timeout of 100ms then the operation can take up to 109 minutes.

The timeout generated by quick-check is meant to be the total timeout, so divide it by the
number of read/write calls that we are expected to make, to ensure that the overall timeout is not excessive.

If the timeout ends up being too small then we tell quickcheck to generate another configuration
with `assume`.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>