CP-49875: Group the auto_instrumentation spans by module
Adds a parent span for the auto_instrumentation of each module. This
helps to group them together so that they can be hidden more easily in
Jaeger. Ideally we should group all of the auto_instrumentation spans
together, but there is no easy way to do this as there's no way to be
sure of when we're finished importing modules.

Signed-off-by: Steven Woods <steven.woods@citrix.com>