CA-395789: Add polling to cluster health state update
It was observed that the corosync-notifyd might send a notification when
the data shown in corosync-quorumtool has not been updated. This causes
xapi-clusterd to return the out-of-date cluster information to xapi.
Xapi will not have a chance of updating these states again until there
is a further change in the cluster (which may not happen for a long
time).

Now add polling in xapi to update the cluster info every 5 minutes,
making sure that in the worst case, we still get up-to-date information
after a 5-minute delay.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>