clock: Add timezone to all parsed dates
Previously the dates without timezone were assumed to use UTC when being
converted to unix time, but the datatype maintained the lack of timezone when
being printed. This means that the problematic timezoneless timestamps were
kept and could be produced by the hosts.

Since the dates are assumed to be UTC, add the timezone when they are parsed
for the first time. Now their timezone is displayed at all times.

The output of Localtime is kept without a timezone because clients are not
prepared to accept arbitrary timezones in dates. (Both the SDK and XO(Lite))

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>