CP-49148: perfmon.service is not loaded
When testing the feature/py3 branch, xapi failed to restart due to
perfmon.service not loaded.

The root cause is that:
In this PR: https://github.com/xapi-project/xen-api/pull/5767, we moved
perfmon.service from `scripts/Makefile` to `python3/Makefile`
In `scripts/Makefile`, `IPROG` is defined as:
IPROG=./install.sh 755
IDATA=./install.sh 644
While in `python3/Makefile`, `IPROG` is defined as:
IPROG=install -m 755
IDATA=install -m 644

And the purpose of `install.sh` is to replace strings
in *.service like: @OPTDIR@ -> ${OPTDIR}
So in python3/Makefile, we didn't replace these strings, and then led to the error.

Signed-off-by: Stephen Cheng <stephen.cheng@cloud.com>