CP-50448: run the QuickCheck tests in QuickTest
Quicktest runs in Dom0, and the existing quickcheck tests run in the CI.
Some of these test as much the OCaml code, as its interaction with the system (e.g. behaviour of system calls).
So it is better to run these tests both in the CI and in Dom0.

We run these in long mode, to explore more randomly generated scenarios.

The seed can be controlled with QCHECK_SEED environment variable.
Similar to @stresstest it uses a random seed, instead of a fixed seed.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>