tracing: increase the default maximum number of spans in a trace
Increases the default maximum number of spans inside a trace from `1000`
to `2500`.

Now that we instrumented the internal http calls, with all components
enabled, the number of spans inside `xapi` component for a `vm-start`
operations is slightly  greater than `1000`. This causes spans to be
leaked, they are removed from the ongoing span table but never added in
the finished tabled. Therefore, they are lost unless the limit is
change in `/etc/xapi.conf`.

This should fix the issue until we implement a better abstraction for
the span hashtables.

Signed-off-by: Gabriel Buica <danutgabriel.buica@cloud.com>