CP-50426: Factor out module signature prefix
This stylistic change is shorter than alternatives. The actual intention is to
ensure that Merlin doesn't get confused when jumping to definition. The
previous behaviour is that Merlin jumps to the type definition rather than
realising it ought to jump to the implementation being referred to inside the
record.

This commit should be squashed together with what follows.

Signed-off-by: Colin James <colin.barr@cloud.com>