xapi-storage-script: remove complexity from the inotify watcher
The number of actions for the plugin directories is reduced to two, reload all
the directory's contents, or reload a file's contents.

This means that when a file is modified is removed and readded to the plugins,
and if the directory changed in any other way the whole directory is reloaded.

The code also tries to reduce duplication in the plugin watchers as much as
possible.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>