CA-396635: Wait for corosync to update its info
Sometimes it takes the underlying cluster stack (corosync) some time to return
a consistent view of the quorum. For example, it may be that the membership
information correctly reflects the new members after a membership change, while the
quorum field is still out of date. Add a delay here to make sure that the information
from corosync represents a consistent snapshot of the current cluster state.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>