CI: avoid mixing caches from different OSes
This is a bug in setup-ocaml@v3 https://github.com/ocaml/setup-ocaml/issues/839

Work it around by defining our own cache prefix based on runner OS version.

Unfortunately the version itself doesn't seem to be available as a variable in GH actions.
There is 'runner.os', but that is just a generic Linux, there is 'matrix.os', but that is only present when using a matrix,
and there is '..container' which is only present when containers are used.

Use another GH action to determine the version, and now the cache-prefix looks like this:
```
cache-prefix: v3-Ubuntu-22.04
```

Signed-off-by: Edwin Török <edwin.torok@cloud.com>