CA-388210: SMAPIv3 concurrency: turn on concurrent operations by default
It is believed that the deadlocks in SMAPIv3 were caused by VDI.copy operations that attach
the same disk RO multiple times in Dom0.
With the previous commits we now use a unique identifier and spawn a separate qemu-dp process
in the SMAPIv3 plugins, which should prevent the deadlocks and IO errors due to lack of refcounting.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>