Output if parameter is required in JSON backend
This change augments all parameter structures emitted by the JSON backend to
include a "required" member that specifies whether the parameter is
optional or not. Parameters are considered optional in the IDL if a default
value is provided for them.

The simplest example is session.login_with_password. This message permits two
extra parameters, "version" and "originator" - but these are not required to
successfully invoke the function.

Signed-off-by: Colin James <colin.barr@cloud.com>