CP-51042: Raise error in sr-scan when SR.stat finds an unhealthy SR
Previously sr-scan first performs an SR.ls followed by an SR.stat, which
will fail when the SR is not healthy. Now that we included more health
messages in SR.stat, we can give perform SR.stat first and fail the
operation if SR.stat gives unhealthy state, and tell the user about the
unhealthy state.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>