CA-397268: vbd.create failed: The device name is invalid
In the commit 62db5cb, the device name validation was consolidated with
the function which converts name to device number. Particularly, the
function "Device_number.of_string" is used for both the validation and
the conversion.

But one issue was introduced in the validation is that the "None" value
returned from "Device_number.of_string" is considered as valid. This
causes the error "the device name is invalid".

This commit just fixes this issue.

Signed-off-by: Ming Lu <ming.lu@cloud.com>