CP-50054: Add authentication cache options
Adds configurable options "enable-external-authentication-cache" and
"external-authentication-cache-size" to Xapi_globs, which control whether the
cache is enabled and the maximum capacity of the cache, respectively.

Also adds "external-authentication-expiry" option, which permits
configuration of the time-to-live assigned to each cache entry. The global that
it manipulates in Xapi_globs was introduced earlier.

In future, these configuration options may become a pool-wide feature.

Signed-off-by: Colin James <colin.barr@cloud.com>