IH-676: Add help strings to suggested commands
These are cleaned up from the final string before it's autocompleted into the
command line. Stylistically, the auxiliary information is right-padded to the
longest command so that it starts on the same vertical line.
The help string is cut to the first sentence since some of these can be
quite long.

Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>