Cache external authentication results
This change provides the workings required to allow the previously functionalised
section of login_with_password to be bypassed by consulting an authentication
cache.

Before attempting to execute that section of login_with_password (which calls
out to external authentication plugins), a cache is consulted. If a
previously-computed result resides within the cache, that is provided. If there
is no such result present in the cache, or external authentication caching is
disabled, the previous section of code (the "slow path") is invoked.

Signed-off-by: Colin James <colin.barr@cloud.com>