IH-676: Do not suggest already entered parameters
Check if the parameter has already been provided and if so, do not include it
in the list of autocompletions anymore.

This should be safe - xe has other ways of entering parameters multiple times:
with commas, or as parts of records. To verify, cli_operations.ml receives the
command line arguments as an associative list and only contains List.assoc and
the like, no List.find_all (apart from the special case in
5d1601c4c5ba020aa847881ad312d380cf0f1ba4 which detects if a parameter was
entered multiple times and errors out, warning the user - which only underlines
that entering a parameter multiple times is not the desired behaviour, can
confuse things and lead to errors)

Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>