IH-689: record_util tests should accept a more lenient new of_string
Before, tests would expect a new version of record_util to throw an exception
where the old version did. We are not harming anyone by instead starting to
be case-insensitive for new of_string functions - this only expands the
accepted set.

Also reformat the autogeneration script to be more readable.

Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>