IH-689: Make old record_util exception behaviour consistent with the new one
Tests expect not only an exception thrown in the new record_util where it was
thrown in the old one, but for it to be the same type. Corrects an outlier in
the old code that returned something other than Record_failure.

The only usage in xapi_message.ml was guarded by a wildcard try X with _ -> ,
so this does not break anything.

Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>