IH-689: record_util.ml - remove manually-specified vbd_mode and vbd_type
vbd_mode was not testing of_string variants because they were not named
according to the convention and were not picked up by the test-generating
script.

vbd_type was not being tested because it did not have a to_string function,
only an of_string one. An automatically generated to_string function is used in
the test instead.

Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>