quicktest: disable open 1024 fds on startup for now
We've enabled running the unixext tests in quicktest.
They open 1024 file descriptors on startup to check for the absence of select.
And although that works for unixext (it is select-free), it doesn't yet work for the rest
of quicktests on master (the required changes to make it work are on the epoll branch).

Temporarily disable this test and add a note to reenable it on the epoll branch.

Fixes: efcb7af9d9d2 ("CP-50448: run the QuickCheck tests in QuickTest")

Signed-off-by: Edwin Török <edwin.torok@cloud.com>