CP-51042: Introduce new SR.scan2 for SMAPI{v1,v2,v3}
Add a new API SR.scan2 as a toplevel function which will will be
implemented differently on SMAPIv1 and SMAPIv3. But not calling it yet,
as that will be done in the following commit.

On SMAPIv1, this will be implemented as a scan followed by a stat,
which is as implemented before. It is important to keep this ordering
because the storage backend relies on the scan call to do the resizing.

On SMAPIv3, this is implemented as a stat followed by an ls, and the ls
is only called if the SR is healthy, as returned by the stat. This is Ok
since on SMAPIv3 SR.stat does the resizing instead.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>