CA-398438: Signal exit to the watcher thread
This solves the problem where disabling and enabling clustering
sometimes makes the thread not start, because it is still in the middle
of processing, and by the time it reaches the end and reset the
`cluster_change_watcher`, `create_as_necessary` has already be called
with `cluster_change_watcher` equal to true. Therefore set
`cluster_change_watcher` immediately to false as soon as the cluster is
to be destroyed, or the coordinator is to be changed.

Note, however, in the case of coorindator change, the watcher will not be
created until the new coordinator has restarted. This is ok since the new
coordinator will only be picked up on next boot.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>