CP-50614: Tracing: replace locks with atomic operations
Tracing shouldn't use locks, because we use it to find hot spots in the application.
And introducing a single global lock might introduce lock contention that wouldn't otherwise be present.

Use atomic operations on immutable data structures instead.
Although a Map is O(log n), and not O(1) like a hashtable, it doesn't require holding any locks to traverse it,
or update it.
We only need to do an atomic compare-and-set operation once we've finished updating it, and if we raced with anyone
 (unlikely on OCaml4, unless you got interrupted by the tick thread), then try again with a backoff.

We shouldn't hand roll atomic data structures like this, but instead use Saturn (Skip lists), or Kcas (which has a
generic [update] function that implements the above method and also works on OCaml 5).

before: 3827.092437 ns/run (confidence: 4275.705106 to 3550.511099);
after:  2727.247326 ns/run (confidence: 3019.854167 to 2582.316754);

Note: when benchmarking ensure that /sys/devices/system/clocksource/clocksource0/current_clocksource is set to TSC.
If set to Xen, then reading timestamps is very slow.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>