CA-397599 XSI-1704 implement setter for blocked ops manually
Currently the setter for field VM.blocked_oeprations is auto generated.
Implement this explicitly such that we can update allowed operations
which currently is not happening. Allowed operations are used by
XenCenter to disable operations; this has led to operations becoming
unavaliable because allowed operations where not aligned with blocked
operations when blocked operations were updated.

Signed-off-by: Christian Lindig <christian.lindig@cloud.com>