fix(WLS): disable non-root unit test
This fails when running kojienv from WSL2, but works when running kojienv from an actual Linux.
In a WSL2 kojienv /dev/loop0 is world writable (even though the user is not root),
whereas in a real kojienv /dev/loop0 is not wordl writable.

This test was here just to test my assumptions that indeed non-root cannot create block devices,
assumptions which are not true in WSL2, so remove this test.

In Dom0 we run as root always, so this has no effect on the product.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>