CA-399187: Allow gencert to be called without groupid
Last year, an argument to determine the group id of the certificates created
was added. This broke test clients, and there's no need no make it compulsory
since there's a default value: -1.

Modify the binary so it can accept only 3 arguments like before, and change the
help output to mention the group id.

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>