CP-51479: [maintenance]: install SDK files using dune rules
Avoid copying twice, get dune to install the files to the correct destination in one go.

Also add a copy of XE_SR_ERRORCODES.xml, by default 'make install' would look
for this in /opt/xensource, and writing there requires root.
(This can be overriden with `./configure --share`).
Since we are using `dune` to install the files now we need the file to always
be present.

Had to adjust the paths used by the CI.

Uses dune directory targets, and the directory must be entirely under the control of these rules.
There are some static files in autogen/ though, so move the generated ones to autogen-out,
and then use 'cp -r' to copy over the static ones
(there is no builtin dune action for the copy, there are individual copy actions,
or a copy_files rule, but neither is suitable here)

Signed-off-by: Edwin Török <edwin.torok@cloud.com>