[ci]: make concurrency group more unique
We notice that required CI checks cancel themselves even when attempting to merge a single PR at a time.
That is probably because there are CI jobs run on both 'push' and 'merge_group'.

Try to make the concurrency group more unique by adding the github event name to the group key.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>