idl: Make lifecycle explicit for callers of 'names'
This is the only commit in the series that changes the schematest hash, because
previous implementation of names created two fields with different lifecycle
descriptions, and when the callers are changed to provide an explicit lifecycle
argument, the description is the same for both fields. It's been chosen to
provide an empty description instead (since any actual description would be
duplicated for both fields, further confusing matters), except for one case
where it follows the rest of the class with 'lifecycle:removed'.

Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>