CA-399638: The livepatches are absent in the response of /updates
When applying livepatches, for one component, only the latest one will
be applied. This is because the latest livepatch will always roll up all
the previous ones if they are of the same component and base build ID.

The bug to be fixed in this commit is that the previous livepatches with
the same build ID are not returned in the response of the query on
the /updates HTTP endpoint.

Signed-off-by: Ming Lu <ming.lu@cloud.com>