CP-49135: split secret and non-secret Refs
Split Uuidx.all into without_secret and secret parts.
This ensures that all sessions are created only with Uuidx.make_uuid_rnd which is guaranteed to use a CSPRNG.
Currently Uuidx.make uses the same generator, but that will change in a future commit.

Introduce Ref.make_secret, Ref.of_secret_string, and change the other function's signatures so that none of them
return a secret (except Ref.null).

It is important that uuids/refs that are used as a secret/auth token are generated in a cryptographically safe way,
if a user can guess them then they can impersonate an authenticated user.

Ref.of_secret_string bypasses these checks, but it is only used in a limited set of places, where we receive
session ids from an external source (HTTP requests, or CLI parameters), or test code.

No functional change.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>