CA-399757: Add CAS style check for SR scan
SR.scan is currently not an atomic operation, and this has caused
problems as during the scan itself, there might be other calls changing
the state of the database, such as VDI.db_introduce called by SM, if
using SMAPIv1. This will confuse SR.scan as it sees an outdated
snapshot.

The proposed workaround would be add a CAS style check for SR.scan,
which will refuse to update the db if it detects changes. This is still
subject to the TOCTOU problem, but should reduce the racing window.

Signed-off-by: Vincent Liu <shuntian.liu2@cloud.com>