CA-399956 - xe autocompletion: Fix autocompletion for words with given prefix
For unresolved reasons, 'compgen -W' would not work as intended occasionally:

'xe host-param-get uuid=xxx param-na<TAB>'

would complete to:

'xe host-param-get uuid=xxx param-name='

but typing:

'xe host-param-get uuid=xxx param-name=so<TAB>'

would not complete it to:

'xe host-param-get uuid=xxx param-name=software-version'

It's unlikely to be a bug in Bash since I couldn't find anything in the
changelog and upgrading the package did not fix it; nor could I reproduce it in
a different program with the same inputs - it is likely that something in the
environment breaks compgen. Switch to pure Bash string processing instead.

Signed-off-by: Andrii Sultanov <andrii.sultanov@cloud.com>