http-libs: add test about error logging
While this does not exercise the exact error that can happen in long
migrations, it gets logged in a similar way.

There's no easy way to trigger the issue, the best chance is to send a
malformed response to trigger a Parse_error.

I did modify the code in http_client and verified that current code can produce
the logging, with backtraces successfully, when set up properly (like in the
test client)

Signed-off-by: Pau Ruiz Safont <pau.ruizsafont@cloud.com>